add audio

+++ mmt/datasets/pipelines/loading.py
@@ -9,7 +9,7 @@ from ..builder import PIPELINES
 class LoadAnnotations(object):
     def __call__(self, results):
         try:
-            results['audio'] = np.load(results['audio_anns'])
+            results['audio'] = np.load(results['audio_anns']).astype(np.float32)
             results['video'] = np.load(results['video_anns'])
             with open(results['text_anns'], 'r', encoding='utf-8') as f:
                 results['text'] = json.load(f)
